add a session resource and tests
I'm not 100% happy with this resource yet. It's too tied to Kioku at the
moment. Also I think it could provide a better extension API that overriding
the base methods (GET, POST, PUT, DELETE). However I think this is a good
sketch for where we would want a Session resource to go, and this helps all
the current Magpie apps.