Throw a 405 HTTP Method not method.
Resources may not be allowed to run a given method. We want to throw an
HTTP 405 in those cases. This adds a default check to see what methods
we're allowed to execute (by default all of the ones from HTTP 1.1) and
throws a 405 if it's outside of that.

We also added a test for teh Abstract Resource that checks the various
levels of HTTP compliance (200, 501, 405).